Actors: Better collision detection by trimming images
Some of the PNG images used for actors have quite large transparent
sides. Since the collision detection algorithm currently only considers
bounds of the ImageView, no matter if it is transparent, this results in
hitboxes much larger than the visible part of the image.

Trimming the edges of the PNG assets also saves a bit of space.