Screen: Switch contentRoot to StackPane
Having StackPane as content root is more versatile, as it is easy
to achieve absolute positioning in StackPane (just add Pane as a child),
but not easy to achieve relative positioning in a Pane (adding StackPane
as a child does not work, as the StackPane will not stretch to the whole
size of the pane).

Also required switching layers to Panes, since Groups
can't resize to fit parent, they are only as big as the content inside.